Доработал класс Bible

+++ backend/src/main/kotlin/app/biblequote/Api.kt
@@ -1,6 +1,6 @@
 package app.biblequote
 
-import app.biblequote.io.Verse
+import app.biblequote.utils.Verse
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
 
@@ -16,21 +16,33 @@ class Api(private val abbrToBible: Map<String, Bible>) {
     TODO()
   }
 
+  /**
+   * Список доступных книг для выбранного перевода
+   */
   @GetMapping("/api/books")
   fun books(translation: String): List<String> {
     TODO()
   }
 
+  /**
+   * Список глав для выбранной книги определенного перевода
+   */
   @GetMapping("/api/chapters")
-  fun chapters(translation: String, book: String): UShort {
+  fun chapters(translation: String, book: String): Int {
     TODO()
   }
 
+  /**
+   * Список стихов для выбранной главы определенной книги и перевода
+   */
   @GetMapping("/api/verses")
-  fun verses(translation: String, book: String, chapter: UShort): List<String> {
+  fun verses(translation: String, book: String, chapter: Int): List<String> {
     TODO()
   }
 
+  /**
+   * Поиск по выбранному переводу
+   */
   @GetMapping("/api/search")
   fun search(translation: String, query: String): List<Verse> {
     TODO()
